fixes for gcc
Former-commit-id: 13eede6621741dfd1e7907b1e59ab6d6d544cef7

+++ oc3_walker_cart_pusher.cpp
@@ -342,7 +342,7 @@ void CartPusher::save( VariantMap& stream ) const
   stream[ "consumerPos" ] = _d->consumerBuilding->getTile().getIJ();
 
   stream[ "maxDistance" ] = _d->maxDistance;
-  stream[ "reservationID" ] = _d->reservationID;
+  stream[ "reservationID" ] = static_cast<int>(_d->reservationID);
 }
 
 void CartPusher::load( const VariantMap& stream )
